+++ pom.xml
@@ -59,6 +59,45 @@
 				<groupId>org.springframework.boot</groupId>
 				<artifactId>spring-boot-maven-plugin</artifactId>
 			</plugin>
+			<plugin>
+				<groupId>com.akathist.maven.plugins.launch4j</groupId>
+				<artifactId>launch4j-maven-plugin</artifactId>
+				<version>2.5.1</version>
+				<executions>
+					<execution>
+						<id>create-exe</id>
+						<phase>package</phase>
+						<goals>
+							<goal>launch4j</goal>
+						</goals>
+					</execution>
+				</executions>
+				<configuration>
+					<dontWrapJar>false</dontWrapJar>
+					<headerType>console</headerType>
+					<jar>C:\Dev\Projects\xml-translator\target\xml-translator-0.0.1-SNAPSHOT.jar</jar>
+					<outfile>C:\Dev\Delivery\Translator\Translator.exe</outfile>
+					<errTitle>Translator</errTitle>
+					<cmdLine/>
+					<chdir>.</chdir>
+					<priority>normal</priority>
+					<downloadUrl>http://java.com/download</downloadUrl>
+					<supportUrl/>
+					<stayAlive>true</stayAlive>
+					<restartOnCrash>false</restartOnCrash>
+					<manifest/>
+					<icon>C:\Dev\Projects\xml-translator\icon.ico</icon>
+					<jre>
+						<path>jre</path>
+						<requiresJdk>false</requiresJdk>
+						<requires64Bit>false</requires64Bit>
+						<minVersion/>
+						<maxVersion/>
+						<initialHeapSize>1024</initialHeapSize>
+						<maxHeapSize>4096</maxHeapSize>
+					</jre>
+				</configuration>
+			</plugin>
 		</plugins>
 	</build>
